mv UpdateRouteStatus to pkg route
